fix parseDsn()
